+++ src/torch_pointcloud/__init__.py
@@ -2,7 +2,7 @@
 
 from importlib.metadata import version as _version
 
-from . import config, datasets, inferers, layers, losses, models, transforms, utils
+from . import config, datasets, inferers, layers, losses, metrics, models, transforms, utils
 from .models import create_model, list_models, register_model
 
 __version__ = _version("torch_pointcloud")
@@ -16,6 +16,7 @@ __all__ = [
     "layers",
     "list_models",
     "losses",
+    "metrics",
     "models",
     "register_model",
     "transforms",